Bit of cleanup, split gauge into kenki + sen
