+++ src/parser/jobs/sam/modules/index.js
@@ -1,17 +1,17 @@
 import Buffs from './Buffs'
 import Combos from './Combos'
-import Gauge from './Gauge'
 import Higanbana from './Higanbana'
 import Kaiten from './Kaiten'
-//import Meditate from './Meditate' DISABLED BECAUSE I MAKE TRASH CODE
+import Kenki from './Kenki'
+import Sen from './Sen'
 import Speedmod from './Speedmod'
 
 export default [
 	Buffs,
 	Combos,
-	Gauge,
 	Higanbana,
 	Kaiten,
-	//Meditate,
+	Kenki,
+	Sen,
 	Speedmod,
 ]
